Android: rename CoreRepository → AppSettings; move to settings/
CoreRepository was a misleading name — it was a thin DataStore wrapper
holding three local preferences (locationId, currentOrderId,
useDemoLocations), not a Ditto-backed data layer. Sharing the
`core/domain/repository/` package with DittoRepository made the
local-vs-synced boundary invisible.

Renamed to AppSettings under a new top-level `live.ditto.pos.settings`
package, mirroring iOS's `iOS/DittoPOS/Settings/SettingsModel.swift`
which holds the same data under `Settings`. The local/synced split is
now visible at the import line.

Mechanical rename across all 21 references; no behavior change.
