+++ Android/app/src/main/java/live/ditto/pos/core/domain/usecase/GetCurrentLocationUseCase.kt
@@ -3,13 +3,13 @@ package live.ditto.pos.core.domain.usecase
 import kotlinx.coroutines.flow.first
 import live.ditto.pos.core.data.demo.LocationSeed
 import live.ditto.pos.core.data.locations.Location
-import live.ditto.pos.core.domain.repository.CoreRepository
 import live.ditto.pos.core.domain.repository.DittoRepository
 import live.ditto.pos.core.domain.usecase.AppConfigurationStateUseCase.AppConfigurationState
+import live.ditto.pos.settings.AppSettings
 import javax.inject.Inject
 
 class GetCurrentLocationUseCase @Inject constructor(
-    private val coreRepository: CoreRepository,
+    private val appSettings: AppSettings,
     private val dittoRepository: DittoRepository,
     private val appConfigurationStateUseCase: AppConfigurationStateUseCase,
     private val isUsingDemoLocationsUseCase: IsUsingDemoLocationsUseCase
@@ -19,7 +19,7 @@ class GetCurrentLocationUseCase @Inject constructor(
         val appConfigurationState = appConfigurationStateUseCase()
 
         return if (appConfigurationState == AppConfigurationState.VALID) {
-            val locationId = coreRepository.locationId()
+            val locationId = appSettings.locationId()
             if (isUsingDemoLocationsUseCase()) {
                 LocationSeed.demoLocations.find { it.id == locationId }
             } else {
