+++ Android/app/src/main/java/live/ditto/pos/core/domain/usecase/SetCurrentLocationUseCase.kt
@@ -1,15 +1,15 @@
 package live.ditto.pos.core.domain.usecase
 
-import live.ditto.pos.core.domain.repository.CoreRepository
+import live.ditto.pos.settings.AppSettings
 import javax.inject.Inject
 
 class SetCurrentLocationUseCase @Inject constructor(
-    private val coreRepository: CoreRepository
+    private val appSettings: AppSettings
 ) {
 
     // Persist the chosen locationId; DittoRepository observes the same flow
     // and registers the orders + sale_items subscriptions for the new value.
     suspend operator fun invoke(locationId: String) {
-        coreRepository.setLocationId(locationId = locationId)
+        appSettings.setLocationId(locationId = locationId)
     }
 }
