+++ Android/app/src/main/java/live/ditto/pos/core/domain/usecase/UpdateCustomLocationUseCase.kt
@@ -1,12 +1,12 @@
 package live.ditto.pos.core.domain.usecase
 
 import live.ditto.pos.core.data.locations.Location
-import live.ditto.pos.core.domain.repository.CoreRepository
 import live.ditto.pos.core.domain.repository.DittoRepository
+import live.ditto.pos.settings.AppSettings
 import javax.inject.Inject
 
 class UpdateCustomLocationUseCase @Inject constructor(
-    private val coreRepository: CoreRepository,
+    private val appSettings: AppSettings,
     private val dittoRepository: DittoRepository,
     private val setCurrentLocationUseCase: SetCurrentLocationUseCase
 ) {
@@ -17,7 +17,7 @@ class UpdateCustomLocationUseCase @Inject constructor(
             name = locationName
         )
         dittoRepository.insertCustomLocation(location = customLocation)
-        coreRepository.shouldUseDemoLocations(useDemoLocations = false)
+        appSettings.shouldUseDemoLocations(useDemoLocations = false)
         setCurrentLocationUseCase(locationId = customLocation.id)
     }
 }
