+++ Android/app/src/main/java/live/ditto/pos/core/domain/usecase/UseDemoLocationUseCase.kt
@@ -1,9 +1,9 @@
 package live.ditto.pos.core.domain.usecase
 
-import live.ditto.pos.core.domain.repository.CoreRepository
+import live.ditto.pos.settings.AppSettings
 import javax.inject.Inject
 
-class UseDemoLocationUseCase @Inject constructor(private val repository: CoreRepository) {
+class UseDemoLocationUseCase @Inject constructor(private val repository: AppSettings) {
 
     suspend operator fun invoke(shouldUseDemoLocations: Boolean) {
         repository.shouldUseDemoLocations(useDemoLocations = shouldUseDemoLocations)
