+++ Android/app/src/main/java/live/ditto/pos/di/AppModule.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import live.ditto.ditto_wrapper.DittoManager
 import live.ditto.pos.BuildConfig
-import live.ditto.pos.core.domain.repository.CoreRepository
+import live.ditto.pos.settings.AppSettings
 import javax.inject.Singleton
 
 @Module
@@ -52,8 +52,8 @@ internal object AppModule {
     }
 
     @Provides
-    fun provideCoreRepository(@ApplicationContext context: Context): CoreRepository {
-        return CoreRepository(context)
+    fun provideAppSettings(@ApplicationContext context: Context): AppSettings {
+        return AppSettings(context)
     }
 
     @Provides
