+++ Android/app/src/main/java/live/ditto/pos/settings/AppSettings.kt
@@ -1,4 +1,4 @@
-package live.ditto.pos.core.domain.repository
+package live.ditto.pos.settings
 
 import android.content.Context
 import androidx.datastore.core.DataStore
@@ -18,8 +18,14 @@ import javax.inject.Singleton
 
 private const val DATA_STORE_NAME = "settings"
 
+/**
+ * Device-local app preferences backed by Android DataStore. Not synced via
+ * Ditto — that's `DittoRepository`'s job. The two are kept in separate
+ * packages so the boundary between "local-only" and "synced" is visible at
+ * the import line. Mirrors iOS `Settings` (`iOS/DittoPOS/Settings/`).
+ */
 @Singleton
-class CoreRepository @Inject constructor(
+class AppSettings @Inject constructor(
     @ApplicationContext private val context: Context
 ) {
